pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "siphon"
version = "0.1.0"
description = "A configurable, YAML-driven, LLM-powered ETL pipeline"
requires-python = ">=3.11"
dependencies = [
    "xmltodict>=0.13.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.0.0",
    "odfpy>=1.4",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pycountry>=24.6.1",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "aiosqlite>=0.19.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
siphon = "siphon.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["siphon"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
